requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
django-rest-swagger==2.2.0
coreapi==2.3.3
coreschema==0.0.4
openapi-codec==1.3.2
factory-boy==3.2.1
Faker==15.3.4
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
phonenumbers==8.13.5
phonenumberslite==8.13.5
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
